1.58: REFACTORING (AMONG OTHER THINGS: FOR CONSISTENCY WITH chronotext-gtest REPO...)

+++ build-android.sh
@@ -1,27 +1,33 @@
 #!/bin/sh
 
-#
-# FOR $HOST_NUM_CPUS, $HOST_OS, $HOST_ARCH
-#
-. `dirname $0`/build-common.sh
-
-if [ -z $NDK_ROOT ]; then
-  echo "NDK_ROOT MUST BE DEFINED!"
-  echo "e.g. export NDK_ROOT=$HOME/android-ndk"
+if [ -z $NDK_PATH ]; then
+  echo "NDK_PATH MUST BE DEFINED!"
   exit -1  
 fi
 
-# ---
-
-BOOST_DIR="boost"
-
-if [ ! -d $BOOST_DIR ]; then
-  echo "ERROR: boost DIRECTORY NOT FOUND"
-  echo "DID YOU EXECUTE init.sh?"
+if [ ! -d dist ]; then
+  echo "ERROR: dist DIRECTORY NOT FOUND!"
+  echo "DID YOU EXECUTE setup.sh?"
   exit 1
 fi
 
-cd $BOOST_DIR
+HOST_OS=$(uname -s | tr "[:upper:]" "[:lower:]")
+HOST_ARCH=$(uname -m)
+HOST_NUM_CPUS=$(sysctl hw.ncpu | awk '{print $2}')
+
+# ---
+
+GCC_VERSION="4.9"
+ANDROID_ABI="armeabi-v7a"
+ANDROID_PLATFORM="android-16"
+
+TOOLCHAIN_PATH=${NDK_PATH}/toolchains/arm-linux-androideabi-${GCC_VERSION}/prebuilt/${HOST_OS}-${HOST_ARCH}
+
+LIBRARIES="--with-system --with-filesystem --with-iostreams"
+
+# ---
+
+cd dist
 
 rm bjam
 rm b2
@@ -31,7 +37,7 @@ rm -rf bin.v2
 ./bootstrap.sh 2>&1
 
 if [ $? != 0 ]; then
-  echo "ERROR: boostrap FAILED"
+  echo "ERROR: boostrap FAILED!"
   exit 1
 fi
 
@@ -39,36 +45,33 @@ cat ../configs/android.jam >> project-config.jam
 
 # ---
 
-LIBRARIES=" --with-system --with-filesystem --with-iostreams"
+LIB_DIR="../lib/android/${ANDROID_ABI}"
 
-LIB_DIR="../lib/android/armeabi-v7a"
-
-GCC_VERSION=4.9
-ANDROID_PLATFORM=android-16
-
-TOOLCHAIN_BIN=${NDK_ROOT}/toolchains/arm-linux-androideabi-${GCC_VERSION}/prebuilt/${HOST_OS}-${HOST_ARCH}/bin
-
-# ---
-
-export PATH=${TOOLCHAIN_BIN}:${PATH}
-export NDK_ROOT
+export PATH=${TOOLCHAIN_PATH}/bin:${PATH}
+export NDK_PATH
 export GCC_VERSION
 export ANDROID_PLATFORM
 export NO_BZIP2=1
 
-./b2 -q -j${HOST_NUM_CPUS}   \
-target-os=android            \
-toolset=gcc-android          \
-link=static                  \
-variant=release              \
-$LIBRARIES                   \
-stage                        \
+./b2 -q -j${HOST_NUM_CPUS}     \
+  target-os=android            \
+  toolset=gcc-android          \
+  link=static                  \
+  variant=release              \
+  $LIBRARIES                   \
+  stage                        \
+  2>&1
+
+if [ $? != 0 ]; then
+  echo "ERROR: b2 FAILED!"
+  exit 1
+fi
+
+# ---
 
 rm -rf $LIB_DIR
 mkdir -p $LIB_DIR
 mv stage/lib/*.a $LIB_DIR
 
-# ---
-
-echo "\nDONE!"
+echo "DONE!"
 ls -1 ${LIB_DIR}/*.a
