+++ build-emscripten.sh
@@ -1,27 +1,25 @@
 #!/bin/sh
 
-#
-# FOR $HOST_NUM_CPUS
-#
-. `dirname $0`/build-common.sh
-
-if [ -z $EMSCRIPTEN_BIN ]; then
-  echo "EMSCRIPTEN_BIN MUST BE DEFINED!"
-  echo "IT SHOULD POINT TO A FOLDER CONTAINING emcc, emar, emranlib AND emlink"
+if [ -z $EMSCRIPTEN_PATH ]; then
+  echo "EMSCRIPTEN_PATH MUST BE DEFINED!"
   exit -1  
 fi
 
-# ---
-
-BOOST_DIR="boost"
-
-if [ ! -d $BOOST_DIR ]; then
-  echo "ERROR: boost DIRECTORY NOT FOUND"
-  echo "DID YOU EXECUTE init.sh?"
+if [ ! -d dist ]; then
+  echo "ERROR: dist DIRECTORY NOT FOUND!"
+  echo "DID YOU EXECUTE setup.sh?"
   exit 1
 fi
 
-cd $BOOST_DIR
+HOST_NUM_CPUS=$(sysctl hw.ncpu | awk '{print $2}')
+
+# ---
+
+LIBRARIES="--with-system --with-filesystem --with-iostreams"
+
+# ---
+
+cd dist
 
 rm bjam
 rm b2
@@ -31,7 +29,7 @@ rm -rf bin.v2
 ./bootstrap.sh 2>&1
 
 if [ $? != 0 ]; then
-  echo "ERROR: boostrap FAILED"
+  echo "ERROR: boostrap FAILED!"
   exit 1
 fi
 
@@ -39,27 +37,29 @@ cat ../configs/emscripten.jam >> project-config.jam
 
 # ---
 
-LIBRARIES=" --with-system --with-filesystem --with-iostreams"
-
 LIB_DIR="../lib/emscripten"
 
-# ---
-
-export PATH=${EMSCRIPTEN_BIN}:${PATH}
+export PATH=${EMSCRIPTEN_PATH}:${PATH}
 export NO_BZIP2=1
 
-./b2 -q -j${HOST_NUM_CPUS}   \
-toolset=clang-emscripten     \
-link=static                  \
-variant=release              \
-$LIBRARIES                   \
-stage                        \
+./b2 -q -j${HOST_NUM_CPUS}     \
+  toolset=clang-emscripten     \
+  link=static                  \
+  variant=release              \
+  $LIBRARIES                   \
+  stage                        \
+  2>&1
+
+if [ $? != 0 ]; then
+  echo "ERROR: b2 FAILED!"
+  exit 1
+fi
+
+# ---
 
 rm -rf $LIB_DIR
 mkdir -p $LIB_DIR
 mv stage/lib/*.a $LIB_DIR
 
-# ---
-
-echo "\nDONE!"
+echo "DONE!"
 ls -1 ${LIB_DIR}/*.a
