+++ build-ios.sh
@@ -1,19 +1,20 @@
 #!/bin/sh
 
-#
-# FOR $HOST_NUM_CPUS
-#
-. `dirname $0`/build-common.sh
-
-BOOST_DIR="boost"
-
-if [ ! -d $BOOST_DIR ]; then
-  echo "ERROR: boost DIRECTORY NOT FOUND"
-  echo "DID YOU EXECUTE init.sh?"
+if [ ! -d dist ]; then
+  echo "ERROR: dist DIRECTORY NOT FOUND!"
+  echo "DID YOU EXECUTE setup.sh?"
   exit 1
 fi
 
-cd $BOOST_DIR
+HOST_NUM_CPUS=$(sysctl hw.ncpu | awk '{print $2}')
+
+# ---
+
+LIBRARIES="--with-system --with-filesystem --with-iostreams"
+
+# ---
+
+cd dist
 
 rm bjam
 rm b2
@@ -23,7 +24,7 @@ rm -rf bin.v2
 ./bootstrap.sh 2>&1
 
 if [ $? != 0 ]; then
-  echo "ERROR: boostrap FAILED"
+  echo "ERROR: boostrap FAILED!"
   exit 1
 fi
 
@@ -31,32 +32,43 @@ cat ../configs/ios.jam >> project-config.jam
 
 # ---
 
-LIBRARIES=" --with-system --with-filesystem --with-iostreams"
+### IPHONE-OS ###
 
 LIB_DIR_1="../lib/ios"
-LIB_DIR_2="../lib/ios-sim"
 
-# ---
+./b2 -q -j${HOST_NUM_CPUS}     \
+  toolset=clang-ios            \
+  link=static                  \
+  variant=release              \
+  $LIBRARIES                   \
+  stage                        \
+  2>&1
 
-./b2 -q -j${HOST_NUM_CPUS}   \
-toolset=clang-ios            \
-link=static                  \
-variant=release              \
-$LIBRARIES                   \
-stage                        \
+if [ $? != 0 ]; then
+  echo "ERROR: b2 FAILED!"
+  exit 1
+fi
 
 rm -rf $LIB_DIR_1
 mkdir -p $LIB_DIR_1
 mv stage/lib/*.a $LIB_DIR_1
 
-# ---
+### IPHONE-SIMULATOR ###
 
-./b2 -q -j${HOST_NUM_CPUS}   \
-toolset=clang-ios_sim        \
-link=static                  \
-variant=release              \
-$LIBRARIES                   \
-stage                        \
+LIB_DIR_2="../lib/ios-sim"
+
+./b2 -q -j${HOST_NUM_CPUS}     \
+  toolset=clang-ios_sim        \
+  link=static                  \
+  variant=release              \
+  $LIBRARIES                   \
+  stage                        \
+  2>&1
+
+if [ $? != 0 ]; then
+  echo "ERROR: b2 FAILED!"
+  exit 1
+fi
 
 rm -rf $LIB_DIR_2
 mkdir -p $LIB_DIR_2
@@ -64,6 +76,6 @@ mv stage/lib/*.a $LIB_DIR_2
 
 # ---
 
-echo "\nDONE!"
+echo "DONE!"
 ls -1 ${LIB_DIR_1}/*.a
 ls -1 ${LIB_DIR_2}/*.a
